+++ CHANGELOG.md
@@ -12,10 +12,12 @@ and this project adheres to
 - Introduce fhpc_addresses extra variable, a hash with containers as keys and
   the list of IP addresses as values.
 - Possibility to run command with SSH paramiko library in addition to ssh binary
-  executable. 
+  executable.
 - cli:
   - Support for tags to filter deployed configuration tasks.
   - Report cluster status in JSON format with --json option.
+  - Introduce fhpc-emulate-slurmu command to emulate random usage of Slurm
+    cluster.
 - conf:
   - Optional support of Rackslab developement deb repositories, disabled by
     default.